refactor: extract step-type dispatch from executor/core into dispatch modules
Extract the step-type-specific match arms from execute_core() into
focused dispatch modules:
- dispatch/prompt.rs: template resolution, runner invocation, TurnEntry
- dispatch/context.rs: shell context step execution
- dispatch/sub_pipeline.rs: sub-pipeline recursion and depth guard
- dispatch/mod.rs: re-exports

execute_core() becomes a thin router delegating to dispatch modules.
The on_result evaluation loop stays in core.rs (post-dispatch, all types).

No behavior changes. All existing tests pass unchanged.
