+++ ail-core/src/executor/core.rs
@@ -7,21 +7,14 @@
 
 #![allow(clippy::result_large_err)]
 
-use std::time::SystemTime;
-
-use crate::config::domain::{
-    ActionKind, Condition, ContextSource, ResultAction, StepBody, MAX_SUB_PIPELINE_DEPTH,
-};
+use crate::config::domain::{ActionKind, Condition, ContextSource, ResultAction, StepBody};
 use crate::error::AilError;
 use crate::runner::{InvokeOptions, RunResult, Runner};
-use crate::session::{Session, TurnEntry};
-use crate::template;
+use crate::session::Session;
 
+use super::dispatch;
 use super::events::ExecuteOutcome;
-use super::helpers::{
-    build_step_runner_box, build_tool_policy, evaluate_on_result, resolve_effective_runner_name,
-    resolve_prompt_file, resolve_step_provider, resolve_step_system_prompts, run_shell_command,
-};
+use super::helpers::evaluate_on_result;
 
 // ── Observer trait ────────────────────────────────────────────────────────────
 
@@ -159,112 +152,6 @@ impl StepObserver for NullObserver {
     }
 }
 
-// ── Sub-pipeline execution ────────────────────────────────────────────────────
-
-/// Load and execute a sub-pipeline, returning a `TurnEntry` for the calling step.
-///
-/// The `path_template` may contain `{{ variable }}` syntax (SPEC §11); it is resolved
-/// against `session` before the file is loaded. The sub-pipeline runs in isolation:
-/// a fresh `Session` is created with the parent's `last_response` as its invocation prompt.
-/// The child's final step response becomes the returned entry's `response` field.
-///
-/// `depth` guards against infinite recursion; exceeding `MAX_SUB_PIPELINE_DEPTH` aborts.
-///
-/// **Note:** sub-pipelines always execute in headless mode, even when the parent pipeline
-/// is running in controlled mode. Sub-pipeline streaming events and HITL gates are not
-/// propagated to the parent's TUI. This is a known limitation (v0.2).
-pub(super) fn execute_sub_pipeline(
-    path_template: &str,
-    prompt_override: Option<&str>,
-    step_id: &str,
-    session: &mut Session,
-    runner: &dyn Runner,
-    depth: usize,
-    base_dir: Option<&std::path::Path>,
-) -> Result<TurnEntry, AilError> {
-    if depth >= MAX_SUB_PIPELINE_DEPTH {
-        return Err(AilError::PipelineAborted {
-            detail: format!(
-                "Step '{step_id}' would exceed the maximum sub-pipeline nesting depth \
-                 of {MAX_SUB_PIPELINE_DEPTH}"
-            ),
-            context: Some(crate::error::ErrorContext::for_step(
-                &session.run_id,
-                step_id,
-            )),
-        });
-    }
-
-    // Resolve template variables in the path (SPEC §11).
-    let resolved_path = template::resolve(path_template, session)
-        .map_err(|e| e.with_step_context(&session.run_id, step_id))?;
-
-    // Resolve ./relative and ../relative paths against the parent pipeline's directory (SPEC §9).
-    let path_buf = if let (true, Some(base)) = (
-        resolved_path.starts_with("./") || resolved_path.starts_with("../"),
-        base_dir,
-    ) {
-        base.join(&resolved_path)
-    } else {
-        std::path::PathBuf::from(&resolved_path)
-    };
-
-    let sub_pipeline = crate::config::load(path_buf.as_path())
-        .map_err(|e| e.with_step_context(&session.run_id, step_id))?;
-
-    // The sub-pipeline's invocation prompt: use the explicit override when provided
-    // (template-resolved against the parent session), otherwise fall back to the
-    // parent's most recent response (SPEC §9).
-    let invocation_prompt = if let Some(override_template) = prompt_override {
-        template::resolve(override_template, session)
-            .map_err(|e| e.with_step_context(&session.run_id, step_id))?
-    } else {
-        session
-            .turn_log
-            .last_response()
-            .unwrap_or(&session.invocation_prompt)
-            .to_string()
-    };
-
-    let mut child_session = crate::session::Session::new(sub_pipeline, invocation_prompt);
-    child_session.cli_provider = session.cli_provider.clone();
-
-    tracing::info!(
-        run_id = %session.run_id,
-        step_id = %step_id,
-        sub_pipeline = %resolved_path,
-        depth,
-        "executing sub-pipeline"
-    );
-
-    execute_core(&mut child_session, runner, &mut NullObserver, depth + 1)?;
-
-    let response = child_session
-        .turn_log
-        .last_response()
-        .unwrap_or("")
-        .to_string();
-
-    Ok(TurnEntry {
-        step_id: step_id.to_string(),
-        prompt: resolved_path,
-        response: Some(response),
-        timestamp: SystemTime::now(),
-        cost_usd: None,
-        input_tokens: 0,
-        output_tokens: 0,
-        runner_session_id: child_session
-            .turn_log
-            .last_runner_session_id()
-            .map(str::to_string),
-        stdout: None,
-        stderr: None,
-        exit_code: None,
-        thinking: None,
-        tool_events: vec![],
-    })
-}
-
 // ── Core loop ─────────────────────────────────────────────────────────────────
 
 /// Inner execution loop shared by headless and controlled modes.
@@ -327,88 +214,20 @@ pub(super) fn execute_core<O: StepObserver>(
         }
 
         let entry = match &step.body {
-            StepBody::Prompt(template_text) => {
-                let template_text = resolve_prompt_file(template_text, &step_id, pipeline_base_dir)
-                    .inspect_err(|e| observer.on_step_failed(&step_id, e.detail()))?;
-                let resolved = template::resolve(&template_text, session)
-                    .map_err(|e| e.with_step_context(&session.run_id, &step_id))
-                    .inspect_err(|e| observer.on_step_failed(&step_id, e.detail()))?;
-
-                observer.on_prompt_ready(&step_id, step_index, total_steps, &resolved);
-
-                let resume_id = if step.resume {
-                    session
-                        .turn_log
-                        .last_runner_session_id()
-                        .map(|s| s.to_string())
-                } else {
-                    None
-                };
-                session.turn_log.record_step_started(&step_id, &resolved);
-
-                let (resolved_system_prompt, resolved_append_system_prompt) =
-                    resolve_step_system_prompts(step, session, &step_id, pipeline_base_dir)
-                        .inspect_err(|e| observer.on_step_failed(&step_id, e.detail()))?;
-
-                let resolved_provider = resolve_step_provider(session, step);
-                let effective_tools = step
-                    .tools
-                    .as_ref()
-                    .or(session.pipeline.default_tools.as_ref());
-                // Update session.runner_name so {{ session.tool }} reflects the active runner.
-                session.runner_name = resolve_effective_runner_name(step);
-                let step_runner_box = build_step_runner_box(
-                    step,
-                    session.headless,
-                    &session.http_session_store,
-                    &resolved_provider,
-                )
-                .inspect_err(|e| observer.on_step_failed(&step_id, e.detail()))?;
-                let effective_runner: &dyn Runner = step_runner_box
-                    .as_deref()
-                    .map(|b| b as &dyn Runner)
-                    .unwrap_or(runner);
-
-                let extensions = effective_runner.build_extensions(&resolved_provider);
-                let mut options = InvokeOptions {
-                    resume_session_id: resume_id,
-                    tool_policy: build_tool_policy(effective_tools),
-                    model: resolved_provider.model,
-                    extensions,
-                    permission_responder: None,
-                    cancel_token: None,
-                    system_prompt: resolved_system_prompt,
-                    append_system_prompt: resolved_append_system_prompt,
-                };
-                observer.augment_options(&mut options);
-
-                let result = observer
-                    .invoke(effective_runner, &resolved, options)
-                    .inspect_err(|e| observer.on_step_failed(&step_id, e.detail()))?;
-
-                tracing::info!(
-                    run_id = %session.run_id,
-                    step_id = %step_id,
-                    cost_usd = ?result.cost_usd,
-                    "step complete"
-                );
-                observer.on_prompt_completed(&step_id, &result);
-
-                TurnEntry::from_prompt(step_id.clone(), resolved, result)
-            }
+            StepBody::Prompt(template_text) => dispatch::prompt::execute(
+                template_text,
+                step,
+                session,
+                runner,
+                &step_id,
+                step_index,
+                total_steps,
+                pipeline_base_dir,
+                observer,
+            )?,
 
             StepBody::Context(ContextSource::Shell(cmd)) => {
-                session.turn_log.record_step_started(&step_id, cmd);
-                let (stdout, stderr, exit_code) = run_shell_command(&session.run_id, &step_id, cmd)
-                    .inspect_err(|e| observer.on_step_failed(&step_id, e.detail()))?;
-                tracing::info!(
-                    run_id = %session.run_id,
-                    step_id = %step_id,
-                    exit_code,
-                    "context shell step complete"
-                );
-                observer.on_non_prompt_completed(&step_id);
-                TurnEntry::from_context(step_id.clone(), cmd.clone(), stdout, stderr, exit_code)
+                dispatch::context::execute_shell(cmd, session, &step_id, observer)?
             }
 
             StepBody::Action(ActionKind::PauseForHuman) => {
@@ -418,23 +237,16 @@ pub(super) fn execute_core<O: StepObserver>(
             StepBody::SubPipeline {
                 path: path_template,
                 prompt,
-            } => {
-                session
-                    .turn_log
-                    .record_step_started(&step_id, path_template);
-                let entry = execute_sub_pipeline(
-                    path_template,
-                    prompt.as_deref(),
-                    &step_id,
-                    session,
-                    runner,
-                    depth,
-                    pipeline_base_dir,
-                )
-                .inspect_err(|e| observer.on_step_failed(&step_id, e.detail()))?;
-                observer.on_non_prompt_completed(&step_id);
-                entry
-            }
+            } => dispatch::sub_pipeline::execute(
+                path_template,
+                prompt.as_deref(),
+                &step_id,
+                session,
+                runner,
+                depth,
+                pipeline_base_dir,
+                observer,
+            )?,
 
             StepBody::Skill(_) => {
                 let err = AilError::PipelineAborted {
@@ -493,7 +305,7 @@ pub(super) fn execute_core<O: StepObserver>(
                         // addressable as `{{ step.<id>__on_result.response }}` without
                         // shadowing the parent step's own turn log entry (SPEC §11).
                         let on_result_step_id = format!("{step_id}__on_result");
-                        let sub_entry = execute_sub_pipeline(
+                        let sub_entry = dispatch::sub_pipeline::execute_sub_pipeline(
                             path,
                             prompt.as_deref(),
                             &on_result_step_id,
