+++ ail-core/src/executor/dispatch/mod.rs
@@ -1,0 +1,5 @@
+//! Step-type-specific dispatch logic extracted from the core execution loop.
+
+pub(super) mod context;
+pub(super) mod prompt;
+pub(super) mod sub_pipeline;
